src/non_functional_requirements/security.py
```python
from cryptography.fernet import Fernet
from src.utils import get_encryption_key

class Security:
    def __init__(self):
        self.key = get_encryption_key()
        self.cipher_suite = Fernet(self.key)

    def encrypt_data(self, data):
        """Encrypts the provided data using Fernet symmetric encryption."""
        encrypted_data = self.cipher_suite.encrypt(data.encode())
        return encrypted_data

    def decrypt_data(self, encrypted_data):
        """Decrypts the provided data using Fernet symmetric encryption."""
        decrypted_data = self.cipher_suite.decrypt(encrypted_data).decode()
        return decrypted_data

    def ensure_gdpr_compliance(self, user_data):
        """Ensures that the user data is handled in compliance with GDPR regulations."""
        # This is a placeholder. In a real-world application, you would implement
        # actual GDPR compliance checks and measures here.
        return True
```