tests/test_non_functional_requirements/test_security.py
```python
import unittest
from src.non_functional_requirements.security import Security

class TestSecurity(unittest.TestCase):

    def setUp(self):
        self.security = Security()

    def test_encrypt_data(self):
        plain_text = "This is a test email content."
        encrypted_text = self.security.encrypt_data(plain_text)
        self.assertNotEqual(plain_text, encrypted_text, "Encryption failed.")

    def test_decrypt_data(self):
        plain_text = "This is a test email content."
        encrypted_text = self.security.encrypt_data(plain_text)
        decrypted_text = self.security.decrypt_data(encrypted_text)
        self.assertEqual(plain_text, decrypted_text, "Decryption failed.")

    def test_compliance_with_GDPR(self):
        user_data = {
            "user_email": "test@example.com",
            "user_password": "password123",
            "trigger_conditions": {},
            "follow_up_content": "This is a follow-up email."
        }
        is_compliant = self.security.compliance_with_GDPR(user_data)
        self.assertTrue(is_compliant, "GDPR compliance check failed.")

if __name__ == "__main__":
    unittest.main()
```